app/requirements.txt
asgiref==3.4.1
blis==0.7.4
catalogue==1.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cymem==2.0.5
Cython==0.29.24
dartsclone==0.9.0
emoji==1.4.2
fastapi==0.68.0
gensim==3.8.1
ginza==4.0.6
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
ja-ginza==4.0.0
murmurhash==1.0.5
numpy==1.21.1
plac==1.1.3
preshed==3.0.5
pydantic==1.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
smart-open==5.1.0
sortedcontainers==2.1.0
spacy==2.3.7
srsly==1.0.5
starlette==0.14.2
SudachiDict-core==20210802
SudachiPy==0.5.2
thinc==7.4.5
tqdm==4.62.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
wasabi==0.8.2
watchgod==0.7
websockets==9.1
mecab-python3==0.996.6rc2
google-cloud-storage==1.41.1
